configurados prettier y eslint

+++ src/components/DeletePost.jsx
@@ -1,43 +1,37 @@
-import { useContext, useState } from "react";
-import { deletePhotosService } from "../services";
-import { AuthContext } from "../context/AuthContext";
+import { useContext, useState } from 'react';
+import { deletePhotosService } from '../services';
+import { AuthContext } from '../context/AuthContext';
 
 function DeletePost({ photo, removePost }) {
-  const { token,idUser,role } = useContext(AuthContext);
-  const [error, setError] = useState("");
+  const { token, idUser, role } = useContext(AuthContext);
+  const [error, setError] = useState('');
 
   const deletePost = async (id) => {
     try {
-       await deletePhotosService({ id, token });
+      await deletePhotosService({ id, token });
 
       if (removePost) {
         removePost(id);
       }
-      
     } catch (error) {
       setError(error.message);
     }
   };
 
   return (
-    
     <>
-    {(role === "admin" || token && parseInt(photo.idUser) === parseInt(idUser)) && (
-          <button
-           className="deletePostButton"
-             onClick={() => {
-            if (
-               window.confirm(`¿Estás seguro de borrar esta publicación?`)
-            )
-             deletePost(photo.idEntry);
-        }}
-      >
-        ❌
-      </button>
-        )}
-      
+      {(role === 'admin' || (token && parseInt(photo.idUser) === parseInt(idUser))) && (
+        <button
+          className='deletePostButton'
+          onClick={() => {
+            if (window.confirm(`¿Estás seguro de borrar esta publicación?`)) deletePost(photo.idEntry);
+          }}
+        >
+          ❌
+        </button>
+      )}
 
-      {error ? <p className="error-message">{error}</p> : null}
+      {error ? <p className='error-message'>{error}</p> : null}
     </>
   );
 }
